issue#33localStorage non-loggedin users cart

+++ client/features/cart/CartSlice.js
@@ -5,40 +5,49 @@ const TOKEN = "token";
 
 export const fetchCart = createAsyncThunk("fetchOrder_Products", async () => {
   const token = window.localStorage.getItem(TOKEN);
-  try {
+  if(token) {
     const { data } = await axios.get(`/api/order_products/cart`, {
       headers: {
         authorization: token,
       },
     });
     return data;
-  } catch (error) {
-    console.log(error);
+  } else {
+    return JSON.parse(window.localStorage.products)
   }
 });
 
+
 export const addItemToCart = createAsyncThunk(
   "addOrder_Product",
-  async ({ productId }) => {
+  async ({ userId, productId }, {dispatch}) => {
     const token = window.localStorage.getItem(TOKEN);
-    try {
+    if(token) {
       const { data } = await axios.post(
-        `/api/order_products/cart`,
-        {
-          productId,
-        },
-        {
-          headers: {
-            authorization: token,
-          },
-        }
-      );
-      return data;
+                `/api/order_products/cart`,
+                {
+                  productId,
+                },
+                {
+                  headers: {
+                    authorization: token,
+                  },
+                }
+              );
+              return data;
 
     } else  {
       const { payload } = await dispatch(fetchSingleProduct(productId))
 
-		  return payload
+      let product ={}
+      for(let key in payload){
+        product[key]=payload[key]
+      }
+      
+      if (!product['quantityInCart']){
+        product['quantityInCart'] = 1
+      }
+		  return product
 
     }
   }
@@ -46,9 +55,9 @@ export const addItemToCart = createAsyncThunk(
 
 export const incrementItemInCart = createAsyncThunk(
   "incrementOrder_Product",
-  async ({ productId, quantityInCart }) => {
+  async ({ productId, quantityInCart }, {dispatch}) => {
     const token = window.localStorage.getItem(TOKEN);
-    try {
+    if(token) {
       quantityInCart++;
       const { data } = await axios.put(
         `/api/order_products/cart`,
@@ -63,8 +72,13 @@ export const incrementItemInCart = createAsyncThunk(
         }
       );
       return data;
-    } catch (error) {
-      console.log(error);
+    }else{
+      let localArray = JSON.parse(window.localStorage.getItem('products'))
+      for (let i =0;i<localArray.length;i++){
+        if (localArray[i].id===productId){
+          localArray[i].quantityInCart++
+        }}
+    window.localStorage.setItem('products', JSON.stringify(localArray))
     }
   }
 );
@@ -73,7 +87,7 @@ export const decrementItemInCart = createAsyncThunk(
   "decrementOrder_Product",
   async ({ productId, quantityInCart }) => {
     const token = window.localStorage.getItem(TOKEN);
-    try {
+    if(token) {
       quantityInCart--;
       const { data } = await axios.put(
         `/api/order_products/cart`,
@@ -88,8 +102,17 @@ export const decrementItemInCart = createAsyncThunk(
         }
       );
       return data;
-    } catch (error) {
-      console.log(error);
+    } else {
+      let localArray = JSON.parse(window.localStorage.getItem('products'))
+      for (let i =0;i<localArray.length;i++){
+        if (localArray[i].id===productId){
+          if(localArray[i].quantityInCart>2){
+          localArray[i].quantityInCart--
+        }else{
+          localArray[i].quantityInCart=1
+        }
+        }}
+    window.localStorage.setItem('products', JSON.stringify(localArray))
     }
   }
 );
@@ -98,15 +121,17 @@ export const removeFromCart = createAsyncThunk(
   "deleteOrder_Product",
   async ({ productId }) => {
     const token = window.localStorage.getItem(TOKEN);
-    try {
+    if(token) {
       await axios.delete(`/api/order_products/${productId}/cart`, {
         headers: {
           authorization: token,
         },
       });
       return productId;
-    } catch (error) {
-      console.log(error);
+    } else {
+      let localArray = JSON.parse(window.localStorage.getItem('products'))
+      localArray = localArray.filter(product => product.id !== productId)
+			window.localStorage.products = JSON.stringify(localArray)
     }
   }
 );
@@ -120,16 +145,40 @@ const cartSlice = createSlice({
   reducers: {},
   extraReducers(builder) {
     builder.addCase(fetchCart.fulfilled, (state, action) => {
-      state.cart = action.payload;
+      const token = window.localStorage.getItem(TOKEN);
+      if(token){
+        state.cart = action.payload;
+      }else{
+        state.cart = JSON.parse(window.localStorage.getItem('products'))
+      }
     });
+
+
     builder.addCase(addItemToCart.fulfilled, (state, action) => {
-      
+      const token = window.localStorage.getItem(TOKEN);
+      if (token){
       state.cart.push(action.payload);
-      localArry.push(action.payload);
-      console.log(localArry)
+      }else{
+      let init = false
+      for (let i =0;i<localArry.length;i++){
+        if (localArry[i].id===action.payload.id){
+          init = true
+          localArry[i].quantityInCart++
+        }
+      }
+      if (!init){
+        localArry.push(action.payload)
+      }
       window.localStorage.setItem('products', JSON.stringify(localArry))
+      }
     });
+
+
+
+
     builder.addCase(incrementItemInCart.fulfilled, (state, action) => {
+      const token = window.localStorage.getItem(TOKEN);
+      if (token){
       state.cart = state.cart.map((item) => {
         let product = item.product;
         if (item.productId === action.payload.productId) {
@@ -137,9 +186,16 @@ const cartSlice = createSlice({
           item["product"] = product;
         }
         return item;
-      });
+      })
+    }
+
     });
+
+
+
     builder.addCase(decrementItemInCart.fulfilled, (state, action) => {
+      const token = window.localStorage.getItem(TOKEN);
+      if (token){
       state.cart = state.cart.map((item) => {
         let product = item.product;
         if (item.productId === action.payload.productId) {
@@ -148,11 +204,18 @@ const cartSlice = createSlice({
         }
         return item;
       });
+    }
     });
+
+
+
     builder.addCase(removeFromCart.fulfilled, (state, action) => {
+      const token = window.localStorage.getItem(TOKEN);
+      if(token){
       state.cart = state.cart.filter(
         (product) => product.productId !== action.meta.arg.productId
       );
+     }
     });
   },
 });
